server/requirements.txt
Django>=3.0,<3.1
psycopg2==2.8.4
bikram==2.1.0
djangorestframework==3.11.0
django-cors-headers==3.3.0
